1.用户查询添加机构

+++ spring-support-common-starter/src/main/java/com/chua/starter/common/support/logger/SysLoggerInfo.java
@@ -61,7 +61,7 @@ public class SysLoggerInfo extends ApplicationEvent {
     /**
      * 耗时(s)
      */
-    private Double logCost;
+    private Long logCost;
     /**
      * 請求地址
      */
